c267: api/client.ts split slice 2 — extract EDA family (#441 2.4)
Continues the c261 incremental split. EDA family (ClassEntry,
ClassMeanSpectrum, ScenePerScene) extracted to api/eda.ts (56 LOC)
with a new edaApi.perScene() runtime method using the shared
_http.ts primitives.

client.ts changes:
- 28 lines of type declarations removed (now in eda.ts).
- client.ts: 1241 -> 1216 lines.
- Type re-exports preserve all existing imports.

The HidsagEda type stays in client.ts for now since it depends on a
long chain of HIDSAG-specific types (HidsagDistribution,
HidsagMethodAggregate, HidsagBlock, ...) — a follow-on cycle will
extract the HIDSAG family as its own slice.

Verification:
- pnpm typecheck -> clean
- pnpm test -> 30 passed
- pnpm build -> clean

Cumulative since c261: client.ts 1392 -> 1216 LOC (-12.6%).
